pyproject.toml
[project]
name = "crow-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "kangminsung",email = "nonasking@gmail.com"}
]
requires-python = ">=3.13,<4.0"
dependencies = [
    "django (>=6.0.2,<7.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "django-environ (>=0.13.0,<0.14.0)",
    "black (>=26.1.0,<27.0.0)",
    "isort (>=8.0.0,<9.0.0)",
    "commitizen (>=4.13.8,<5.0.0)",
    "notion-client (>=3.0.0,<4.0.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-django (>=4.12.0,<5.0.0)",
    "drf-spectacular (>=0.29.0,<0.30.0)",
    "requests (>=2.32.5,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.commitizen]
name = "cz_conventional_commits"
version_provider = "pep621"
tag_format = "v$version"
